# Much better polygon cover algorithm