Makes polygon cover algorithm pretty much as fast and simple as it can get. Closes #58. 

- ~20-25% faster according to benchmarks (update: up to 3x faster)
- fixes cover bugs (see unnecessary tiles removed in the out files)
- much much simpler, tons of code removed

@morganherlocker @yhahn you'll like this!