@yhahn 2015-09-03
@mourner awesome, RIP local min/max ascii art diagrams :joy: