@mourner 2015-09-04
Pushed 2 more changes:

- faster merge algorithm that leads to ~20% faster polygon cover when it's merged
- fix multipoint not being merged