@mourner 2015-09-04
Pushed another perf improvement that leads to up to 3x gain in some polygon cases (where there are a lot of fills):

```bash
# before
scan russia polygon - z6 - z6 x 4,849 ops/sec ±2.24% (23 runs sampled)
scan russia polygon - z8 - z8 x 429 ops/sec ±3.56% (22 runs sampled)
scan russia polygon - z10 - z10 x 25.72 ops/sec ±1.72% (13 runs sampled)
scan russia polygon multizoom - z0 - z9 x 80.18 ops/sec ±2.82% (18 runs sampled)

# after
scan russia polygon - z6 - z6 x 6,493 ops/sec ±2.12% (23 runs sampled)
scan russia polygon - z8 - z8 x 1,001 ops/sec ±2.20% (23 runs sampled)
scan russia polygon - z10 - z10 x 81.71 ops/sec ±8.53% (18 runs sampled)
scan russia polygon multizoom - z0 - z9 x 113 ops/sec ±3.30% (19 runs sampled)
```