@morganherlocker 2015-09-09
:racehorse: published in v3.0.1!