@mourner 2015-09-09
Ha, you're a risky guy for publishing a patch version after half of the code rewritten from scratch :) Thanks Morgan!