@morganherlocker 2015-09-09
@mourner I thought about doing a major, but decided a patch made more sense. This did not change anything functionally (that we know of), other than bugs. If we find any functional - but valid - differences, we can roll this back and publish a major version bump. 

That's semver for you ¯\/(°_o)\/¯